pyproject.toml
[project]
name = "lang-proj"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "faiss-cpu>=1.13.2",
    "ipykernel>=7.2.0",
    "langchain>=1.2.14",
    "langchain-anthropic>=1.4.0",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.23",
    "langchain-groq>=1.1.2",
    "langchain-openai>=1.1.12",
    "pdfminer>=20191125",
    "pypdf>=6.9.2",
    "python-dotenv>=1.2.2",
    "unstructured>=0.22.6",
]
